HZN-1371: Introduce SimpleHealthCheck and consolidate existing simple health checks
